pyproject.toml
[project]
name = "task1-advance-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.1",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.2",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.52.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
